Updates to form selection page.
