+++ jha/src/App.vue
@@ -34,10 +34,20 @@ import {
   MODULE_NAME as appModule,
   SET_SHOW_MOBILE_STEPPER_DETAILS,
 } from '@/store/modules/app-module';
+import {
+  MODULE_NAME as enrolmentModule,
+  SET_APPLICATION_UUID,
+} from '@/store/modules/enrolment-module';
 import pageStateService from '@/services/page-state-service';
 import { scrollTo } from '@/helpers/scroll';
-import { enrolmentRoutes } from './router/routes';
+import {
+  commonRoutes,
+  enrolmentRoutes,
+} from './router/routes';
 import { Wormhole } from 'portal-vue';
+import { v4 as uuidv4 } from 'uuid';
+import spaEnvService from '@/services/spa-env-service';
+import logService from '@/services/log-service';
 
 export default {
   name: 'App',
@@ -54,6 +64,26 @@ export default {
   },
   created() {
     document.title = this.pageTitle;
+
+    this.applicationUuid = uuidv4();
+    this.$store.dispatch(enrolmentModule + '/' + SET_APPLICATION_UUID, this.applicationUuid);
+
+    // Load environment variables, and route to maintenance page.
+    spaEnvService.loadEnvs()
+      .then(() => {
+        if (spaEnvService.values && spaEnvService.values.SPA_ENV_OOP_MAINTENANCE_FLAG === 'true') {
+          const toPath = commonRoutes.MAINTENANCE_PAGE.path;
+          pageStateService.setPageComplete(toPath);
+          pageStateService.visitPage(toPath);
+          this.$router.push(toPath);
+        }
+      })
+      .catch((error) => {
+        logService.logError(this.applicationUuid, {
+          event: 'HTTP error getting values from spa-env-server',
+          status: error.response.status,
+        });
+      });
   },
   computed: {
     stepRoutes() {
